host/requirements.txt
pyserial>=3.5
numpy>=1.21
sounddevice>=0.4
